add Uid back to json, use old json script to run dev
