+++ packages/slice-machine/lib/models/common/CustomType/index.ts
@@ -1,16 +1,12 @@
 import { Tab, TabAsArray, TabAsObject } from './tab'
 import { SliceZone, SliceZoneAsArray } from './sliceZone'
-import { Field, FieldType } from './fields'
+import { Field } from './fields'
 import { UIDField } from '../widgets/UID/type'
-import { AsObject } from '../widgets/Group/type'
+import { UID } from './uid'
 
 export type ObjectTabs = {
   [key: string]: TabAsObject
 }
-
-function isTabAsObject(obj: object): obj is TabAsObject {
-  return 'key' in obj && 'value' in obj
-} 
 
 export type ArrayTabs = ReadonlyArray<TabAsArray>
 
@@ -19,91 +15,65 @@ export interface SeoTab {
   description: string;
 }
 
-interface CustomTypeJson {
-  [key: string]: {
-    [fieldId: string]: Field | SliceZone
-  }
-}
 export interface CustomTypeJsonModel {
   id: string;
   status: boolean;
   repeatable: boolean;
   label: string;
-  json: CustomTypeJson
+  json: {
+    [key: string]: {
+      [fieldId: string]: Field | SliceZone
+    }
+  }
 };
+
 export interface CustomType<T extends ObjectTabs | ArrayTabs> {
   id: string;
   status: boolean;
   repeatable: boolean;
   label: string;
   tabs: T;
-  uid: UIDField | null;
+  uid?: UIDField;
   previewUrl?: string;
-}
-
-function extractUidFromTab(value: AsObject): UIDField | null {
-  return Object
-    .entries(value)
-    .reduce((f: UIDField | null, [_, field]: [string, Field | SliceZone]) => {
-      return  field.type === FieldType.UID
-        ? field as UIDField
-        : f
-    }, null)
-}
-
-function extractUidFromTabs(ct: CustomType<ObjectTabs> | CustomTypeJson): UIDField | null {
-  return Object
-      .entries(ct.tabs)
-      .reduce((uid: UIDField | null, [_, tab]: [string, TabAsObject | AsObject]) => {
-        if (uid) return uid
-        const uidField = extractUidFromTab(
-          isTabAsObject(tab) ? tab.value : tab
-        )
-        return uidField || null
-  }, null)
 }
 
 export const CustomType = {
   toArray(ct: CustomType<ObjectTabs>): CustomType<ArrayTabs> {
-    const uid = extractUidFromTabs(ct)
+    const uid = UID.extractUidFromTabs(ct.tabs)
 
     return {
       ...ct,
-      uid,
+      ... uid ? { uid } : {},
       tabs: Object.entries(ct.tabs).map(([key, value]) => Tab.toArray(key, value)),
     }
   },
   toObject(ct: CustomType<ArrayTabs>): CustomType<ObjectTabs> {
-    return {
-      ...ct,
-      tabs: ct.tabs.reduce((acc, tab) => {
+    const { tabs, uid, ...rest } = ct
+
+    const objectTabs: ObjectTabs =
+      ct
+      .tabs
+      .reduce((acc, tab) => {
         return {
           ...acc,
           [tab.key]: Tab.toObject(tab)
         }
-      }, {}),
+      }, {})
+
+    const updatedTabs = uid
+      ? UID.addUidToFirstTab(objectTabs, uid)
+      : objectTabs
+
+    return {
+      ...rest,
+      tabs: updatedTabs
     }
   },
   toJsonModel(ct: CustomType<ObjectTabs>): CustomTypeJsonModel {
     const { tabs, previewUrl, uid, ...rest } = ct
 
     const updatedTabs = uid
-      ? (() => {
-          const firstTabName: string | undefined = Object.keys(tabs)[0]
-          if (!firstTabName) throw new Error("Wrong custom type with no default tab.")
-
-          const { [firstTabName]: firstTab, ...rest } = tabs
-          
-          const updatedFirstTab: TabAsObject = {
-            key: firstTabName,
-            value: {
-              "uid": uid,
-              ...firstTab.value
-            }
-          }
-
-          return { [firstTabName]: updatedFirstTab, ...rest }
-        })()
+      ? UID.addUidToFirstTab(tabs, uid)
       : tabs
 
     return {
@@ -119,18 +89,21 @@ export const CustomType = {
   },
   fromJsonModel(key: string, ct: CustomTypeJsonModel): CustomType<ObjectTabs> {
     const { json, ...rest } = ct
-    const uid: UIDField | null = extractUidFromTabs(json)
 
+    const tabs: ObjectTabs = Object.entries(json).reduce((acc, [key, value]) => {
+      return {
+        ...acc,
+        [key]: { key, value }
+      }
+    }, {})
+
+    const uid: UIDField | null = UID.extractUidFromTabs(tabs)
+    
     return {
       ...rest,
       id: key,
-      uid: uid,
-      tabs: Object.entries(json).reduce((acc, [key, value]) => {
-        return {
-          ...acc,
-          [key]: { key, value }
-        }
-      }, {}),
+      tabs,
+      ... uid ? { uid } : {},
     }
   },
   getSliceZones(ct: CustomType<ArrayTabs>): ReadonlyArray<SliceZoneAsArray | null> {
