+++ packages/slice-machine/package.json
@@ -15,7 +15,8 @@
     "postinstall": "node postinstall",
     "prepublishOnly": "yarn bundle",
     "serve": "node -r esm ./build/server/src/index.js",
-    "dev-server": "cross-env PORT=9999 CWD=./tests/project nodemon -e ts,js --watch server --ignore '*.test.ts' --exec ts-node server/src/index.js",
+    "watch-server": "tsc -w --project server/tsconfig.json",
+    "dev-server": "cross-env PORT=9999 CWD=./tests/project node build/server/src/index.js",
     "build": "yarn build-server && yarn build-next-app",
     "build-server": "tsc --project server/tsconfig.json",
     "build-next-app": "cross-env NODE_OPTIONS=--max_old_space_size=8192 next build",
